backend/python-ml/requirements.txt
numpy==1.24.3
torch==2.0.1
transformers==4.30.2
sentence-transformers==2.2.2
fastapi==0.95.2
uvicorn==0.22.0
python-multipart==0.0.6
pillow==9.5.0
scikit-learn==1.2.2 